Initial commit of a simple RSS reader
This is a very early prototype of an RSS reader in Java. The main point
of this is to learn basic Android development and try remember the Java
I learnt at University.

The functionality of the project so fat is limitted to getting a hard
coded list of RSS feeds, parsing them into basic channel and item
objects and displying them in a list. Currently you can only view the
items in each list and not actually content of the entries. Admittedly
this makes for a terrible RSS reader, however, this is still a very
early prototype.

Future development work includes:
 - Making content viewable.
 - Adding methods to add and remove RSS feeds
 - Marking items as read or unread.
 - Adding the ability to flag or star items
 - Storing RSS feeds and content locally for offline access
 - Only viewing items that need to be read
Or basically the core functionality of any RSS reader!