RssFeedParser: append hash of the feed url when saving as a file
Some feeds are given have the same title despite having different
contents (e.g. all MangaDex RSS feeds have the same name "MangaDex
RSS"). Since the feeds are saved to disk in files based on their title
this means the feeds with the same title are saved into the same
location overwriting eachother.

Now a hash of the URL is appended to the title ensure that channels with
the same title are stored separately. Note, this does ignore the fringe
case where several channels with the same title are stored at the same
URL. However, this is a very unique edge-case and should not need
supporting.